refactor: clean up error handling and improve error types
- Consolidate error handling patterns across modules
- Remove redundant error variants
- Improve error messages for better debugging
