+++ cda-main/src/main.rs
@@ -12,20 +12,14 @@
  */
 
 use cda_core::DiagServiceResponseStruct;
-use cda_interfaces::{DiagServiceError, DoipGatewaySetupError, dlt_ctx};
+use cda_interfaces::dlt_ctx;
 use cda_plugin_security::{DefaultSecurityPlugin, DefaultSecurityPluginData};
-use cda_tracing::TracingSetupError;
 use clap::Parser;
 use futures::future::FutureExt;
 use opensovd_cda_lib::{
+    AppError,
     config::configfile::{ConfigSanity, Configuration},
     setup_tracing, shutdown_signal,
-};
-use thiserror::Error;
-
-use crate::AppError::{
-    ConfigurationError, ConnectionError, DataError, InitializationFailed, NotFound, ResourceError,
-    RuntimeError, ServerError,
 };
 
 #[derive(Parser, Debug)]
@@ -67,96 +61,6 @@ struct AppArgs {
     log_file_name: Option<String>,
 }
 
-#[derive(Error, Debug)]
-pub enum AppError {
-    #[error("Initialization failed `{0}`")]
-    InitializationFailed(String),
-    #[error("Resource error: `{0}`")]
-    ResourceError(String),
-    #[error("Connection error `{0}`")]
-    ConnectionError(String),
-    #[error("Configuration error `{0}`")]
-    ConfigurationError(String),
-    #[error("Data error `{0}`")]
-    DataError(String),
-    #[error("Error during execution `{0}`")]
-    RuntimeError(String),
-    #[error("Not found: `{0}`")]
-    NotFound(String),
-    #[error("Server error: `{0}`")]
-    ServerError(String),
-}
-
-impl From<TracingSetupError> for AppError {
-    fn from(value: TracingSetupError) -> Self {
-        match value {
-            TracingSetupError::ResourceCreationFailed(_) => ResourceError(value.to_string()),
-            TracingSetupError::SubscriberInitializationFailed(_) => {
-                InitializationFailed(value.to_string())
-            }
-        }
-    }
-}
-
-impl From<DoipGatewaySetupError> for AppError {
-    fn from(value: DoipGatewaySetupError) -> Self {
-        match value {
-            DoipGatewaySetupError::InvalidAddress(_) => ConnectionError(value.to_string()),
-
-            DoipGatewaySetupError::SocketCreationFailed(_)
-            | DoipGatewaySetupError::PortBindFailed(_) => InitializationFailed(value.to_string()),
-
-            DoipGatewaySetupError::InvalidConfiguration(_) => ConfigurationError(value.to_string()),
-
-            DoipGatewaySetupError::ResourceError(_) => ResourceError(value.to_string()),
-
-            DoipGatewaySetupError::ServerError(_) => ServerError(value.to_string()),
-        }
-    }
-}
-
-impl From<DiagServiceError> for AppError {
-    fn from(value: DiagServiceError) -> Self {
-        match value {
-            DiagServiceError::RequestNotSupported(_)
-            | DiagServiceError::BadPayload(_)
-            | DiagServiceError::ConnectionClosed
-            | DiagServiceError::UnexpectedResponse(_)
-            | DiagServiceError::EcuOffline(_)
-            | DiagServiceError::NoResponse(_)
-            | DiagServiceError::SendFailed(_)
-            | DiagServiceError::InvalidAddress(_)
-            | DiagServiceError::InvalidRequest(_)
-            | DiagServiceError::Timeout => ConnectionError(value.to_string()),
-
-            DiagServiceError::ParameterConversionError(_)
-            | DiagServiceError::UnknownOperation
-            | DiagServiceError::UdsLookupError(_)
-            | DiagServiceError::VariantDetectionError(_)
-            | DiagServiceError::AccessDenied(_)
-            | DiagServiceError::InvalidSession(_)
-            | DiagServiceError::Nack(_) => RuntimeError(value.to_string()),
-
-            DiagServiceError::InvalidSecurityPlugin => ConfigurationError(value.to_string()),
-
-            DiagServiceError::ResourceError(_) => ResourceError(value.to_string()),
-
-            DiagServiceError::NotFound(Some(_)) => NotFound(value.to_string()),
-
-            DiagServiceError::NotFound(None) => NotFound("Resource could not be found.".to_owned()),
-
-            DiagServiceError::DataError(_)
-            | DiagServiceError::InvalidDatabase(_)
-            | DiagServiceError::DatabaseEntryNotFound(_)
-            | DiagServiceError::NotEnoughData { .. } => DataError(value.to_string()),
-
-            DiagServiceError::SetupError(_) | DiagServiceError::ConfigurationError(_) => {
-                InitializationFailed(value.to_string())
-            }
-        }
-    }
-}
-
 #[tokio::main]
 #[tracing::instrument(
     fields(
@@ -194,11 +98,7 @@ async fn main() -> Result<(), AppError> {
         &config,
         clonable_shutdown_signal.clone(),
     )
-    .await
-    .map_err(|e| {
-        let err: AppError = e.into();
-        err
-    })?;
+    .await?;
 
     cda_sovd::add_vehicle_routes::<DiagServiceResponseStruct, _, _, DefaultSecurityPlugin>(
         &dynamic_router,
